# openapi.yaml
openapi: 3.0.3
info:
  title: Defellix API
  description: >
    API Documentation for the Defellix Decentralized Freelancer Trust Platform.
    This API gateway routes requests to underlying microservices: Auth, User, Contract, and Blockchain.
  version: 1.0.0
servers:
  - url: https://api.defellix.com
    description: Production API via Cloudflare Tunnels
  - url: http://localhost:8080
    description: Local development via Nginx API Gateway
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Enter your JWT token in the format **Bearer &lt;token&gt;**
  schemas:
...
    SubmissionRequest:
      type: object
      required: [submitted_data, description]
      properties:
        status:
          type: string
          enum: [draft, pending_review]
        milestone_id:
          type: integer
          description: Optional. Set to link this submission to a specific milestone.
        submitted_data:
          type: object
          description: Dynamic JSON mapping fulfilling the exact 'submission_criteria' previously defined.
        description:
          type: string
          description: Free-form description of the delivered work and any links to files/repo.
    UpdateSubmissionRequest:
      type: object
      properties:
        status:
          type: string
          enum: [draft, pending_review]
        submitted_data:
          type: object
          description: Dynamic JSON mapping fulfilling the exact 'submission_criteria' previously defined.
        description:
          type: string
          description: Free-form description of the delivered work and any links to files/repo.
    ClientReviewRequest:
      type: object
      required: [action]
      properties:
        action:
          type: string
          enum: [accept, revision]
          description: Use "accept" to approve and finalize the work. Use "revision" to reject it back to the freelancer with mandatory feedback.
        comment:
          type: string
          description: Required if action is 'revision'. Optional feedback if 'accept'.
paths:
...
  /api/v1/contracts/{id}/submissions:
    get:
      tags: [Submission]
      summary: List submissions for a contract
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: List of submissions returned.
    post:
      tags: [Submission]
      summary: Submit work against a signed contract
      description: Freelancers submit structured JSON (`submitted_data`) mapped to the explicit `submission_criteria` in the Milestone. Defaults to `draft` state unless strictly specified as `pending_review`.
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubmissionRequest'
      responses:
        '201':
          description: Work submitted and client notified via email.
  /api/v1/contracts/{id}/submissions/{sub_id}:
    put:
      tags: [Submission]
      summary: Update a drafted or revision-requested submission
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: sub_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateSubmissionRequest'
      responses:
        '200':
          description: Submission updated successfully.
...
  /api/v1/public/contracts/{token}/submissions/{sub_id}/review:
    post:
      tags: [Submission, Client]
      summary: Review submitted work (Client)
      description: Client can accept or reject the deliverable. If accepted (action=accept), the contract completes, Reputation Points are sent, and the contract is saved to the public profile. If rejected (action=revision), the comment is permanently injected into the JSONB revision_history audit trail.
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
        - name: sub_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClientReviewRequest'
      responses:
        '200':
          description: Review processed.

  # ==========================================
  # BLOCKCHAIN SERVICE
  # ==========================================
...